MirOTR: Libgcrypt, 64bit fix, patch submitted to Libgcrypt devs
git-svn-id: http://svn.miranda-ng.org/main/trunk@12509 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/MirOTR/Libgcrypt/random/rndw32.c
@@ -888,7 +888,7 @@ _gcry_rndw32_gather_random_fast (void (*add)(const void*, size_t,
   {
     HANDLE handle;
     FILETIME creationTime, exitTime, kernelTime, userTime;
-    DWORD minimumWorkingSetSize, maximumWorkingSetSize;
+    SIZE_T minimumWorkingSetSize, maximumWorkingSetSize;
 
     handle = GetCurrentThread ();
     GetThreadTimes (handle, &creationTime, &exitTime,
